# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: HelixAgent AI Platform API
  description: |
    Comprehensive API for AI-powered completions with multi-provider support,
    ensemble reasoning, Cognee AI enhancement, and advanced monitoring capabilities.
  version: 1.0.0
  contact:
    name: HelixAgent Support
    email: support@helixagent.ai
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.helixagent.ai/v1
    description: Production server
  - url: http://localhost:8080/v1
    description: Development server
security:
  - bearerAuth: []
paths:
...
  /v1/auth/register:
    post:
      summary: Register new user
      description: Creates a new user account
      operationId: registerUser
      tags:
        - Authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/auth/login:
    post:
      summary: User login
      description: Authenticates user and returns JWT token
      operationId: loginUser
      tags:
        - Authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/auth/refresh:
    post:
      summary: Refresh token
      description: Refreshes the JWT token
      operationId: refreshToken
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/auth/me:
    get:
      summary: Get current user info
      description: Returns information about the authenticated user
      operationId: getCurrentUser
      tags:
        - Authentication
      responses:
        '200':
          description: User info retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/completions:
    post:
      summary: Create completion
      description: Creates a completion using the specified model
      operationId: createCompletion
      tags:
        - Completions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompletionRequest'
      responses:
        '200':
          description: Completion created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompletionResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/chat/completions:
    post:
      summary: Create chat completion
      description: Creates a chat completion with message history
      operationId: createChatCompletion
      tags:
        - Chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatCompletionRequest'
      responses:
        '200':
          description: Chat completion created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatCompletionResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/ensemble/completions:
    post:
      summary: Create ensemble completion
      description: Creates a completion using ensemble of multiple LLM providers
      operationId: createEnsembleCompletion
      tags:
        - Ensemble
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnsembleCompletionRequest'
      responses:
        '200':
          description: Ensemble completion created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnsembleCompletionResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/models/metadata:
    get:
      summary: List model metadata
      description: Returns metadata for all available models from Models.dev
      operationId: listModelMetadata
      tags:
        - Models
      security: []
      parameters:
        - name: capability
          in: query
          description: Filter by capability
          schema:
            type: string
        - name: provider
          in: query
          description: Filter by provider
          schema:
            type: string
      responses:
        '200':
          description: Model metadata list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelMetadataListResponse'
  /v1/models/metadata/{id}:
    get:
      summary: Get model metadata
      description: Returns detailed metadata for a specific model
      operationId: getModelMetadata
      tags:
        - Models
      security: []
      parameters:
        - name: id
          in: path
          required: true
          description: Model ID
          schema:
            type: string
      responses:
        '200':
          description: Model metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelMetadata'
        '404':
          description: Model not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/models/metadata/{id}/benchmarks:
    get:
      summary: Get model benchmarks
      description: Returns benchmark data for a specific model
      operationId: getModelBenchmarks
      tags:
        - Models
      security: []
      parameters:
        - name: id
          in: path
          required: true
          description: Model ID
          schema:
            type: string
      responses:
        '200':
          description: Model benchmarks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelBenchmarks'
        '404':
          description: Model not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/models/metadata/compare:
    get:
      summary: Compare models
      description: Compares multiple models side by side
      operationId: compareModels
      tags:
        - Models
      security: []
      parameters:
        - name: models
          in: query
          description: Comma-separated list of model IDs to compare
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Model comparison
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelComparison'
...
  /v1/admin/models/metadata/refresh:
    post:
      summary: Refresh model metadata
      description: Triggers a refresh of model metadata from Models.dev (admin only)
      operationId: refreshModelMetadata
      tags:
        - Admin
      responses:
        '200':
          description: Refresh initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Refresh initiated
                  taskId:
                    type: string
        '403':
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/admin/models/metadata/refresh/status:
    get:
      summary: Get refresh status
      description: Returns the status of the model metadata refresh task (admin only)
      operationId: getRefreshStatus
      tags:
        - Admin
      responses:
        '200':
          description: Refresh status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshStatus'
        '403':
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/sessions:
    post:
      summary: Create a new session
      description: Creates a new conversation session
      operationId: createSession
      tags:
        - Sessions
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                config:
                  type: object
      responses:
        '201':
          description: Session created
    get:
      summary: List all sessions
      description: Returns all active sessions
      operationId: listSessions
      tags:
        - Sessions
      responses:
        '200':
          description: Sessions list
  /v1/sessions/{id}:
    get:
      summary: Get session details
      description: Returns details of a specific session
      operationId: getSession
      tags:
        - Sessions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Session details
    delete:
      summary: Terminate session
      description: Terminates and removes a session
      operationId: terminateSession
      tags:
        - Sessions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Session terminated
...
  /v1/lsp/execute:
    post:
      summary: Execute LSP request
      description: Executes a Language Server Protocol request
      operationId: executeLSPRequest
      tags:
        - LSP
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                method:
                  type: string
                params:
                  type: object
      responses:
        '200':
          description: LSP response
...
  /v1/mcp/tools/call:
    post:
      summary: Call MCP tool
      description: Executes an MCP tool
      operationId: callMCPTool
      tags:
        - MCP
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                arguments:
                  type: object
      responses:
        '200':
          description: Tool execution result
...
  /v1/protocols/execute:
    post:
      summary: Execute protocol request
      description: Executes a protocol-agnostic request
      operationId: executeProtocolRequest
      tags:
        - Protocols
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Protocol response
...
  /v1/protocols/configure:
    post:
      summary: Configure protocols
      description: Updates protocol configurations
      operationId: configureProtocols
      tags:
        - Protocols
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Configuration updated
  /v1/embeddings/generate:
    post:
      summary: Generate embeddings
      description: Generates vector embeddings for text
      operationId: generateEmbeddings
      tags:
        - Embeddings
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                input:
                  type: array
                  items:
                    type: string
                model:
                  type: string
      responses:
        '200':
          description: Embeddings generated
  /v1/embeddings/search:
    post:
      summary: Vector search
      description: Performs similarity search using embeddings
      operationId: vectorSearch
      tags:
        - Embeddings
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                query:
                  type: string
                limit:
                  type: integer
      responses:
        '200':
          description: Search results
  /v1/embeddings/index:
    post:
      summary: Index document
      description: Indexes a document for vector search
      operationId: indexDocument
      tags:
        - Embeddings
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Document indexed
  /v1/embeddings/batch-index:
    post:
      summary: Batch index documents
      description: Indexes multiple documents at once
      operationId: batchIndexDocuments
      tags:
        - Embeddings
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Documents indexed
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token
  schemas:
...
    RegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
...
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        refreshToken:
          type: string
        expiresAt:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/UserInfo'
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
        timestamp:
          type: string
          format: date-time
          description: When the error occurred
...